% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborPlots.R
\name{clusterDistanceBoxplot}
\alias{clusterDistanceBoxplot}
\title{Cluster Distance Boxplot}
\usage{
clusterDistanceBoxplot(
  scHolography.obj,
  annotationToUse = "orig.cluster",
  query.cluster.list,
  reference.cluster,
  palette = "Paired",
  n.neighbor = 30
)
}
\description{
Cluster Distance Boxplot
}
