% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHolography.R
\name{distToscHolography}
\alias{distToscHolography}
\title{Reconstruct scHolography graph from distance matrix}
\usage{
distToscHolography(high.res.sp, low.res.sp, dist.mat)
}
\description{
Reconstruct scHolography graph from distance matrix
}
