% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGeneDynamics.R
\name{DoMultiBarHeatmap}
\alias{DoMultiBarHeatmap}
\title{Do MultiBar Heatmap
This function is adapted from https://github.com/elliefewings/DoMultiBarHeatmap}
\usage{
DoMultiBarHeatmap(
  object,
  features = NULL,
  cells = NULL,
  group.by = "ident",
  additional.group.by = NULL,
  additional.group.sort.by = NULL,
  cols.use = NULL,
  group.bar = TRUE,
  disp.min = -2.5,
  disp.max = NULL,
  slot = "scale.data",
  assay = NULL,
  label = TRUE,
  size = 5.5,
  hjust = 0,
  angle = 45,
  raster = TRUE,
  draw.lines = TRUE,
  lines.width = NULL,
  group.bar.height = 0.02,
  combine = TRUE,
  palette = c("Paired", "RdYlBu"),
  disPlot = F
)
}
\description{
Do MultiBar Heatmap
This function is adapted from https://github.com/elliefewings/DoMultiBarHeatmap
}
