% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGeneDynamics.R
\name{expressionByDistPlot}
\alias{expressionByDistPlot}
\title{Expression by Distance Plot}
\usage{
expressionByDistPlot(
  scHolography.obj,
  query.cluster,
  ref.cluster,
  geneOI,
  annotationToUse = "orig.cluster",
  assayToUse = "SCT",
  pal = "Paired",
  n.neighbor = 30
)
}
\description{
Expression by Distance Plot
}
