% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGeneDynamics.R
\name{findDriverGene}
\alias{findDriverGene}
\title{Find Space Driver Gene}
\usage{
findDriverGene(
  scHolography.obj,
  query.cluster,
  ref.cluster,
  k1 = 1.96,
  k2 = 1.96,
  annotationToUse = "orig.cluster",
  n.closeNeighbor = 30,
  assayToUse = "SCT",
  bandwidth = NULL
)
}
\description{
Find Space Driver Gene
}
