% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGeneDynamics.R
\name{findGeneSpatialDynamics}
\alias{findGeneSpatialDynamics}
\title{Find Gene Spatial Dynamics}
\usage{
findGeneSpatialDynamics(
  scHolography.obj,
  query.cluster,
  ref.cluster,
  annotationToUse = "orig.cluster",
  assayToUse = "SCT",
  n.neighbor = 30,
  n.background = 100,
  back.sig = F
)
}
\description{
Find Gene Spatial Dynamics
}
