% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGeneDynamics.R
\name{findSpatialNeighborhood}
\alias{findSpatialNeighborhood}
\title{Find Spatial Neighborhood}
\usage{
findSpatialNeighborhood(
  scHolography.obj,
  annotationToUse,
  query.cluster,
  nclus = NULL,
  pal = "Set3"
)
}
\description{
Find Spatial Neighborhood
}
