% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGeneDynamics.R
\name{findSpatialNeighborhood_new}
\alias{findSpatialNeighborhood_new}
\title{Find Spatial Neighborhood}
\usage{
findSpatialNeighborhood_new(
  scHolography.obj,
  annotationToUse = "orig.ident",
  query.cluster,
  orig.assay = "RNA",
  nDims = 10,
  resolution = 0.2,
  pal = "Paired",
  nNeighborhood = NULL,
  seed = 60611
)
}
\arguments{
\item{scHolography.obj}{scHolography object list}

\item{annotationToUse}{Which annotation to call identities from. Default is orig.cluster}

\item{query.cluster}{A vector of query identity types}

\item{orig.assay}{Which assay to get feature from. Default is RNA}

\item{nDims}{The number of PC dimensions used to find spatial neighborhood. The default is 10}

\item{resolution}{The resolution used for finding spatial neighborhood. The default is 0.2}

\item{pal}{Color palette to use for scHolographyPlot. Default is Paired}

\item{nNeighborhood}{The number of of neighborhood to do k.mean clustering}
}
\description{
Find Spatial Neighborhood
}
