% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHolography.R
\name{interpretDecomp}
\alias{interpretDecomp}
\title{Stable Matching Neighbor Analysis}
\usage{
interpretDecomp(est.array, nslot, is3D = T)
}
\arguments{
\item{est.array}{Distance matrix between individual cells/pixels}

\item{nslot}{Maximum number of neighbors to find for stable matching}

\item{is3D}{If est.array is a 3D tensor. Default is True}
}
\description{
Stable Matching Neighbor Analysis
}
