% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborPlots.R
\name{neighborCompositionPlot}
\alias{neighborCompositionPlot}
\title{Layer Composition Plot from Given Celltypes}
\usage{
neighborCompositionPlot(
  scHolography.obj,
  annotationToUse = "orig.cluster",
  query.cluster,
  palette = "Paired"
)
}
\description{
Layer Composition Plot from Given Celltypes
}
