% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborPlots.R
\name{neighborMarkerPlot}
\alias{neighborMarkerPlot}
\title{Find Neighbor Determining Markers}
\usage{
neighborMarkerPlot(
  scHolography.obj,
  annotationToUse = "orig.cluster",
  query.cluster,
  target.cluster,
  palette = "Paired",
  assayToUse = "SCT",
  cutoff = 1,
  n.DEG = 10,
  heatmap.pal = "viridis"
)
}
\description{
Find Neighbor Determining Markers
}
