% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborPlots.R
\name{neighborSankeyPlot}
\alias{neighborSankeyPlot}
\title{Neighbor Sankey Plot}
\usage{
neighborSankeyPlot(
  scHolography.obj,
  annotationToUse = "orig.cluster",
  query.cluster,
  palette = "Paired",
  font.size = 12
)
}
\description{
Neighbor Sankey Plot
}
