% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGeneDynamics.R
\name{relativeSpatialAnalysis}
\alias{relativeSpatialAnalysis}
\title{Relative Spatial Analysis}
\usage{
relativeSpatialAnalysis(
  scHolography.obj,
  query.cluster,
  ref.cluster,
  plotByDist = F,
  geneOI = NULL,
  annotationToUse = "orig.cluster",
  assayToUse = "SCT",
  quant = 0.1,
  nCperL = NULL,
  nL = NULL,
  heatmapGp = NULL,
  pal = "Paired",
  n.neighbor = 30
)
}
\description{
Relative Spatial Analysis
}
