% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{scHolography3dDistributionPlot}
\alias{scHolography3dDistributionPlot}
\title{3D Cell State Distribution Plot}
\usage{
scHolography3dDistributionPlot(
  scHolography.obj,
  dimToPlot,
  color.by = "celltype",
  palette = "Paired",
  transparency = 0.7
)
}
\description{
3D Cell State Distribution Plot
}
