% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{scHolographyConnectionPlot}
\alias{scHolographyConnectionPlot}
\title{Cell State Connection Plot in 3D}
\usage{
scHolographyConnectionPlot(
  scHolography.obj,
  color.by = "orig.cluster",
  palette = "Paired",
  seed = 60611,
  n.closeNeighbor = 30,
  n.mutualNeighbor = NA
)
}
\description{
Cell State Connection Plot in 3D
}
