% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapToSpatialData.R
\name{scHolographyMapToPixel}
\alias{scHolographyMapToPixel}
\title{scHolography 2D Single Cell Spatial Mapping to ST Data}
\usage{
scHolographyMapToPixel(scHolography.obj, nCperP = NULL, seed = 60611)
}
\description{
scHolography 2D Single Cell Spatial Mapping to ST Data
}
