% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{scHolographyNeighborCompPlot}
\alias{scHolographyNeighborCompPlot}
\title{Spatial Neighborhood Composition Plot}
\usage{
scHolographyNeighborCompPlot(
  scHolography.obj,
  annotationToUse,
  query.cluster = NULL,
  pal = "Paired",
  color = NULL,
  row.group.level = NULL
)
}
\description{
Spatial Neighborhood Composition Plot
}
