% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrativeLearning.R
\name{scHolographySpatialIntegration}
\alias{scHolographySpatialIntegration}
\title{scHolography Joint Learning from Multiple ST references}
\usage{
scHolographySpatialIntegration(spatial.int.list, nslot = 30, seed = 60611)
}
\description{
scHolography Joint Learning from Multiple ST references
}
