% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapToSpatialData.R
\name{scHolographyToPixelPlot}
\alias{scHolographyToPixelPlot}
\title{Visualization of scHolography 2D Single Cell Spatial Mapping}
\usage{
scHolographyToPixelPlot(
  scHolography.obj,
  color.by = "orig.cluster",
  cutoff = NULL,
  assayToUse = "SCT",
  feature = NULL,
  palette = "Paired",
  dot.size = 5,
  cells = NULL,
  highlight = NULL
)
}
\description{
Visualization of scHolography 2D Single Cell Spatial Mapping
}
