% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHolography.R
\name{seuratSCT}
\alias{seuratSCT}
\title{Data Normalization}
\usage{
seuratSCT(seuratObj, ndims)
}
\arguments{
\item{seuratObj}{Seurat Object}

\item{ndims}{Number of PCs used}
}
\description{
Data Normalization
}
