% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGeneDynamics.R
\name{spatialDynamicsFeaturePlot}
\alias{spatialDynamicsFeaturePlot}
\title{Spatial Dynamics Heatmap}
\usage{
spatialDynamicsFeaturePlot(
  scHolography.obj,
  query.cluster,
  ref.cluster,
  geneOI = NULL,
  annotationToUse = "orig.cluster",
  assayToUse = "RNA",
  heatmapGp = NULL,
  pal = "Paired",
  multi.pal = "RdYlBu",
  n.neighbor = 30,
  heatmap.pal = "viridis"
)
}
\arguments{
\item{scHolography.obj}{scHolography object list}

\item{query.cluster}{A vector of query identity types}

\item{ref.cluster}{A vector of reference identity types}

\item{geneOI}{Gene of interest to plot for heatmaps}

\item{annotationToUse}{Which annotation to call identities from. Default is orig.cluster}

\item{assayToUse}{Which assay to get feature from. Default is SCT}

\item{heatmapGp}{Which annotation to use for heatmap column grouping. Default is NULL}

\item{pal}{Color palette to use for coloring. Default is Paired}

\item{n.neighbor}{Number of nearest cells to use to define distance. Default is 30}

\item{heatmap.pal}{Color palette to use for heatmap feature coloring. Default is viridis. Other options are "rdbu", "magma", and "rdbu_1"}
}
\description{
Spatial Dynamics Heatmap
}
