% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHolography.R
\name{dataAlign}
\alias{dataAlign}
\title{Data Integration}
\usage{
dataAlign(
  low.res.sp,
  high.res.sp,
  stProcessed = F,
  scProcessed = F,
  nFeatureToUse = 3000,
  whichReference = 2,
  nPCtoUse = 32
)
}
\description{
Data Integration
}
