% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGeneDynamics.R
\name{expressionByDistPlot}
\alias{expressionByDistPlot}
\title{Expression by Distance Plot}
\usage{
expressionByDistPlot(
  scHolography.obj,
  query.cluster,
  ref.cluster,
  geneOI,
  annotationToUse = "orig.cluster",
  assayToUse = "SCT",
  pal = "Paired",
  n.neighbor = 30
)
}
\arguments{
\item{scHolography.obj}{scHolography object list}

\item{query.cluster}{A vector of query identity types}

\item{ref.cluster}{A vector of reference identity types}

\item{geneOI}{Gene of interest to plot for heatmaps}

\item{annotationToUse}{Which annotation to call identities from. Default is orig.cluster}

\item{assayToUse}{Which assay to get feature from. Default is SCT}

\item{pal}{Color palette to use for coloring. Default is Paired}

\item{n.neighbor}{Number of nearest cells to use to define distance. Default is 30}
}
\description{
Expression by Distance Plot
}
