% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHolography.R
\name{findDistance}
\alias{findDistance}
\title{Find scHolography Inferred Distance between Cell Clusters}
\usage{
findDistance(
  scHolography.obj,
  query.cluster,
  ref.cluster,
  annotationToUse = "orig.cluster",
  n.neighbor = 30
)
}
\description{
Find scHolography Inferred Distance between Cell Clusters
}
