% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGeneDynamics.R
\name{findSpatialNeighborhood}
\alias{findSpatialNeighborhood}
\title{Find Spatial Neighborhood}
\usage{
findSpatialNeighborhood(
  scHolography.obj,
  annotationToUse,
  query.cluster,
  nclus = NULL,
  pal = "Set3"
)
}
\arguments{
\item{scHolography.obj}{scHolography object list}

\item{annotationToUse}{Which annotation to call identities from. Default is orig.cluster}

\item{query.cluster}{A vector of query identity types}

\item{nclus}{Number of spatial neighborhood to cluster. Default is NULL (the number will be decided by silhouette coefficients)}

\item{pal}{Color palette to use for coloring. Default is Set3}
}
\description{
Find Spatial Neighborhood
}
