% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborPlots.R
\name{neighborCompositionPlot}
\alias{neighborCompositionPlot}
\title{Layer Composition Plot from Given Celltypes}
\usage{
neighborCompositionPlot(
  scHolography.obj,
  annotationToUse = "orig.cluster",
  query.cluster,
  palette = "Paired"
)
}
\arguments{
\item{scHolography.obj}{scHolography object list}

\item{annotationToUse}{Which annotation to call identities from. Default is orig.cluster}

\item{query.cluster}{A vector of query identity types}

\item{palette}{Color palette to use for coloring. Default is Paired}
}
\description{
Layer Composition Plot from Given Celltypes
}
