% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGeneDynamics.R
\name{relativeSpatialAnalysis}
\alias{relativeSpatialAnalysis}
\title{Relative Spatial Analysis}
\usage{
relativeSpatialAnalysis(
  scHolography.obj,
  query.cluster,
  ref.cluster,
  plotByDist = F,
  geneOI = NULL,
  annotationToUse = "orig.cluster",
  assayToUse = "SCT",
  quant.left = 0.1,
  quant.right = 0.9,
  nCperL = NULL,
  nL = NULL,
  heatmapGp = NULL,
  pal = "Paired",
  multi.pal = "RdYlBu",
  n.neighbor = 30,
  extreme.comp = F,
  heatmap.pal = "viridis"
)
}
\arguments{
\item{scHolography.obj}{scHolography object list}

\item{query.cluster}{A vector of query identity types}

\item{ref.cluster}{A vector of reference identity types}

\item{plotByDist}{Whether order the heatmap by the SMN distance. Default is False}

\item{geneOI}{Gene of interest to plot for heatmaps}

\item{annotationToUse}{Which annotation to call identities from. Default is orig.cluster}

\item{assayToUse}{Which assay to get feature from. Default is SCT}

\item{quant.left}{Quantile cutoff for the left side. Default is 0.1}

\item{quant.right}{Quantile cutoff for the right side. Default is 0.9}

\item{nCperL}{Manually define number of cell per layer for analysis. Default is NULL}

\item{nL}{Manually define number of cell layers for analysis. Default is NULL}

\item{heatmapGp}{Which annotation to use for heatmap column grouping. Default is NULL}

\item{pal}{Color palette to use for coloring. Default is Paired}

\item{n.neighbor}{Number of nearest cells to use to define distance. Default is 30}

\item{extreme.comp}{Whether to consider only the two extreme (proximal and distal) ends for analysis. Default is False}

\item{heatmap.pal}{Color palette to use for heatmap feature coloring. Default is viridis. Other options are "rdbu", "magma", and "rdbu_1"}
}
\description{
Relative Spatial Analysis
}
