% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{scHolographyPlot}
\alias{scHolographyPlot}
\title{scHolography 3D Tissue Reconstruction Visualization}
\usage{
scHolographyPlot(
  scHolography.obj,
  dim = 3,
  cells = NULL,
  feature = NULL,
  cutoff = NULL,
  color.by = "orig.cluster",
  dot.size = 5,
  palette = "Paired",
  highlight = NULL,
  assayToUse = "SCT"
)
}
\description{
scHolography 3D Tissue Reconstruction Visualization
}
