% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGeneDynamics.R
\name{spatialDynamicsFeaturePlot}
\alias{spatialDynamicsFeaturePlot}
\title{Spatial Dynamics Heatmap}
\usage{
spatialDynamicsFeaturePlot(
  scHolography.obj,
  query.cluster,
  ref.cluster,
  geneOI = NULL,
  annotationToUse = "orig.cluster",
  assayToUse = "RNA",
  heatmapGp = NULL,
  pal = "Paired",
  n.neighbor = 30,
  heatmap.pal = "viridis"
)
}
\description{
Spatial Dynamics Heatmap
}
