% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHolography.R
\name{trainHolography}
\alias{trainHolography}
\title{Single cell 3D reconstruction}
\usage{
trainHolography(
  sp.integrated,
  vSeed = 60611,
  n.repeat = 30,
  n.slot = 30,
  n.pcUse = 32,
  n.pcOut = 32
)
}
\description{
Single cell 3D reconstruction
}
